chore: last fixes
