+++ src/decorators/check_metamask_notification.py
@@ -1,15 +1,5 @@
-import pygetwindow
 import time
-from src.decorators.force_full_screen import forceFullScreenForThis
-
-def maximizeMetamaskNotification():
-    title = 'MetaMask Notification'
-    time.sleep(7)
-    windows = pygetwindow.getWindowsWithTitle(title)
-    if len(windows) > 0:
-        current_window = windows[0]
-        current_window.activate()
-        current_window.maximize()
+from src.bot.action import maximizeMetamaskNotification
 
 def checkMetamaskNotification(fn):
     def exec(*args, **kwargs):
